server/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
boto3==1.28.85
botocore==1.31.85
celery==5.3.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
defusedxml==0.8.0rc2
dj-rest-auth==4.0.1
Django==4.2.21
django-cors-headers==4.1.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
idna==3.10
jmespath==1.0.1
kombu==5.5.3
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.5
Pillow==9.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
pydub==0.25.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
redis==4.6.0
requests==2.32.3
requests-oauthlib==2.0.0
requirements.py==1.0.1
s3transfer==0.7.0
six==1.17.0
social-auth-app-django==5.2.0
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
